refactor: split useSimpleChat tests
