+++ tests/useSimpleChat/init.test.tsx
@@ -1,0 +1,12 @@
+import { renderHook, waitFor } from '@testing-library/react';
+import { describe, expect, it } from 'vitest';
+import { createWrapper } from './setup';
+import { useSimpleChat } from '../../app/hooks/useSimpleChat';
+
+describe('useSimpleChat', () => {
+  it('initializes with expected mock messages', async () => {
+    const wrapper = createWrapper();
+    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
+    await waitFor(() => expect(result.current.docs.length).toBeGreaterThan(0));
+  });
+});
