+++ tests/useSimpleChat/setup.tsx
@@ -1,10 +1,10 @@
-import { act, cleanup, renderHook, waitFor } from '@testing-library/react';
+import { cleanup } from '@testing-library/react';
 import type { ReactNode } from 'react';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { AuthProvider } from '../app/contexts/AuthContext';
+import { afterEach, beforeEach, vi } from 'vitest';
+import { AuthProvider } from '../../app/contexts/AuthContext';
 
 // Mock AuthContext to avoid state updates during tests
-vi.mock('../app/contexts/AuthContext', () => {
+vi.mock('../../app/contexts/AuthContext', () => {
   return {
     AuthProvider: ({ children }: { children: ReactNode }) => <>{children}</>,
     useAuth: () => ({
@@ -17,10 +17,8 @@ vi.mock('../app/contexts/AuthContext', () => {
     }),
   };
 });
-import { useSession } from '../app/hooks/useSession'; // Import the actual hook for vi.mocked
-import { useSimpleChat } from '../app/hooks/useSimpleChat';
-import type { AiChatMessage, ChatMessage } from '../app/types/chat';
-import { parseContent } from '../app/utils/segmentParser';
+import type { AiChatMessage, ChatMessage } from '../../app/types/chat';
+import { parseContent } from '../../app/utils/segmentParser';
 
 // Helper function to convert chunks into SSE format
 function formatAsSSE(chunks: string[]): string[] {
@@ -48,22 +46,22 @@ function formatAsSSE(chunks: string[]): string[] {
 }
 
 // Mock the prompts module
-vi.mock('../app/prompts', () => ({
+vi.mock('../../app/prompts', () => ({
   makeBaseSystemPrompt: vi.fn().mockResolvedValue('Mocked system prompt'),
 }));
 
 // Mock the provisioning module
-vi.mock('../app/config/provisioning');
+vi.mock('../../app/config/provisioning');
 
 // Import the mocked module
-import { getCredits } from '../app/config/provisioning';
-import { createOrUpdateKeyViaEdgeFunction } from '../app/services/apiKeyService';
+import { getCredits } from '../../app/config/provisioning';
+import { createOrUpdateKeyViaEdgeFunction } from '../../app/services/apiKeyService';
 
 // Mock the apiKeyService module
-vi.mock('../app/services/apiKeyService');
+vi.mock('../../app/services/apiKeyService');
 
 // Mock the env module
-vi.mock('../app/config/env', () => ({
+vi.mock('../../app/config/env', () => ({
   CALLAI_API_KEY: 'mock-callai-api-key-for-testing',
   FIREPROOF_CHAT_HISTORY: 'test-chat-history',
   GA_TRACKING_ID: 'mock-ga-tracking-id',
@@ -160,7 +158,7 @@ const mergeUserMessageImpl = (data: any) => {
 const mockMergeUserMessage = vi.fn(mergeUserMessageImpl);
 
 // Mock the useSession hook
-vi.mock('../app/hooks/useSession', () => {
+vi.mock('../../app/hooks/useSession', () => {
   return {
     useSession: () => {
       // Don't reset here, reset is done in beforeEach
@@ -452,7 +450,7 @@ export default Timer;`,
 });
 
 // Mock the useSessionMessages hook
-vi.mock('../app/hooks/useSessionMessages', () => {
+vi.mock('../../app/hooks/useSessionMessages', () => {
   // Track messages across test runs
   const messagesStore: Record<string, ChatMessage[]> = {};
 
@@ -692,646 +690,76 @@ const createWrapper = () => {
   return ({ children }: { children: ReactNode }) => <AuthProvider>{children}</AuthProvider>;
 };
 
-describe('useSimpleChat', () => {
-  const testJwt =
-    'eyJhbGciOiJub25lIiwidHlwIjoiSldUIn0=.eyJ1c2VySWQiOiJ0ZXN0LXVzZXItaWQiLCJleHAiOjI1MzQwMjMwMDc5OX0=.';
+const testJwt =
+  'eyJhbGciOiJub25lIiwidHlwIjoiSldUIn0=.eyJ1c2VySWQiOiJ0ZXN0LXVzZXItaWQiLCJleHAiOjI1MzQwMjMwMDc5OX0=.';
+beforeEach(() => {
+  vi.mocked(createOrUpdateKeyViaEdgeFunction).mockImplementation(async () => {
+    return {
+      key: 'mock-api-key-for-testing',
+      hash: 'mock-hash',
+      name: 'Mock Session Key',
+      label: 'mock-session',
+      limit: 0.01,
+      disabled: false,
+      usage: 0,
+      created_at: new Date().toISOString(),
+      updated_at: new Date().toISOString(),
+    };
+  });
 
-  beforeEach(() => {
-    // Mock createOrUpdateKeyViaEdgeFunction to ensure it returns the correct structure
-    vi.mocked(createOrUpdateKeyViaEdgeFunction).mockImplementation(async () => {
-      return {
-        key: 'mock-api-key-for-testing',
-        hash: 'mock-hash',
-        name: 'Mock Session Key',
-        label: 'mock-session',
-        limit: 0.01,
-        disabled: false,
-        usage: 0,
-        created_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      };
-    });
+  vi.mocked(getCredits).mockImplementation(async () => {
+    return {
+      available: 0.005,
+      usage: 0.005,
+      limit: 0.01,
+    };
+  });
 
-    // Mock the getCredits function
-    vi.mocked(getCredits).mockImplementation(async () => {
-      return {
-        available: 0.005,
-        usage: 0.005,
-        limit: 0.01,
-      };
-    });
-
-    // Mock window.fetch
-    vi.spyOn(window, 'fetch').mockImplementation(async () => {
-      // Mock response with a readable stream
-      const stream = new ReadableStream({
-        start(controller) {
-          controller.enqueue(new TextEncoder().encode('This is a test response'));
-          controller.close();
-        },
-      });
-
-      return {
-        ok: true,
-        body: stream,
-        status: 200,
-        statusText: 'OK',
-        headers: new Headers(),
-      } as Response;
-    });
-
-    // Mock ScrollIntoView
-    Element.prototype.scrollIntoView = vi.fn();
-
-    // Mock environment variables
-    vi.stubEnv('VITE_CALLAI_API_KEY', 'test-api-key');
-
-    // Mock import.meta.env.MODE for testing
-    vi.stubGlobal('import', {
-      meta: {
-        env: {
-          MODE: 'test',
-          VITE_CALLAI_API_KEY: 'test-api-key',
-        },
+  vi.spyOn(window, 'fetch').mockImplementation(async () => {
+    const stream = new ReadableStream({
+      start(controller) {
+        controller.enqueue(new TextEncoder().encode('This is a test response'));
+        controller.close();
       },
     });
 
-    // Reset the mock state before each test
-    resetMockState();
-
-    // Add localStorage mock
-    vi.spyOn(Storage.prototype, 'getItem');
-    localStorage.getItem = vi.fn((key) => {
-      if (key === 'auth_token') return testJwt;
-      return null;
-    });
+    return {
+      ok: true,
+      body: stream,
+      status: 200,
+      statusText: 'OK',
+      headers: new Headers(),
+    } as Response;
   });
 
-  afterEach(() => {
-    cleanup();
-    vi.restoreAllMocks();
-    vi.unstubAllEnvs();
-    vi.unstubAllGlobals();
-    localStorage.clear();
-  });
+  Element.prototype.scrollIntoView = vi.fn();
 
-  it('initializes with expected mock messages', async () => {
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-    await waitFor(() => expect(result.current.docs.length).toBeGreaterThan(0));
-    // ... assertions ...
-  });
+  vi.stubEnv('VITE_CALLAI_API_KEY', 'test-api-key');
 
-  it.skip('updates input value', async () => {
-    const wrapper = createWrapper();
-    renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-    // ... assertions ...
-  });
-
-  it.skip('sends a message and receives a response', async () => {
-    // Create a mock fetch that returns a stream with an AI response
-    const mockFetch = vi.fn().mockImplementation(async () => {
-      const encoder = new TextEncoder();
-      const stream = new ReadableStream({
-        start(controller) {
-          // Simulate a streaming response
-          const chunks = ['Hello', '! How can I help ', 'you today?'];
-
-          // Format chunks as SSE and send them
-          const sseChunks = formatAsSSE(chunks);
-          sseChunks.forEach((chunk) => {
-            controller.enqueue(encoder.encode(chunk));
-          });
-
-          controller.close();
-        },
-      });
-
-      return {
-        ok: true,
-        body: stream,
-        status: 200,
-        statusText: 'OK',
-        headers: new Headers(),
-      } as Response;
-    });
-
-    window.fetch = mockFetch;
-
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-
-    // Set input and verify it was set
-    await act(async () => {
-      result.current.setInput('Hello, AI!');
-    });
-
-    // Check if the mock function was called
-    expect(mockMergeUserMessage).toHaveBeenCalledWith({ text: 'Hello, AI!' });
-
-    // Force a re-render before waiting
-    act(() => {
-      result.current.setInput('force-refresh');
-    });
-
-    // Wait for the input state to update
-    await waitFor(() => {
-      expect(result.current.input).toBe('Hello, AI!');
-    });
-
-    // When we call sendMessage, it should use mergeUserMessage to update and submit the message
-    await act(async () => {
-      await result.current.sendMessage();
-    });
-
-    // Should have messages from our mock
-    expect(result.current.docs.length).toBeGreaterThan(0);
-
-    // Verify that the mock fetch was called, indicating that the API call was attempted
-    expect(mockFetch).toHaveBeenCalled();
-  });
-
-  it('correctly parses markdown and code segments', async () => {
-    // Create a mock fetch that just returns an empty response
-    const mockFetch = vi.fn().mockImplementation(async () => {
-      const encoder = new TextEncoder();
-      const stream = new ReadableStream({
-        start(controller) {
-          controller.enqueue(
-            encoder.encode('data: {"choices":[{"delta":{"content":""},"finish_reason":null}]}\n\n')
-          );
-          controller.close();
-        },
-      });
-
-      return {
-        ok: true,
-        body: stream,
-        status: 200,
-        statusText: 'OK',
-        headers: new Headers(),
-      } as Response;
-    });
-
-    window.fetch = mockFetch;
-
-    // Mock renderHook to inject our custom messages
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-
-    // For this test, we are going to manually construct the messages array
-    // This bypasses all the mock complexity
-    const codeContent = `function HelloWorld() {
-  return <div>Hello, World!</div>;
-}
-
-export default HelloWorld;`;
-
-    const mockMessages = [
-      {
-        type: 'user',
-        text: 'Create a React component',
-        timestamp: Date.now() - 1000,
+  vi.stubGlobal('import', {
+    meta: {
+      env: {
+        MODE: 'test',
+        VITE_CALLAI_API_KEY: 'test-api-key',
       },
-      {
-        type: 'ai',
-        text: `Here's a simple React component:
-
-\`\`\`jsx
-${codeContent}
-\`\`\`
-
-You can use this component in your application.`,
-        segments: [
-          {
-            type: 'markdown' as const,
-            content: "Here's a simple React component:",
-          },
-          {
-            type: 'code' as const,
-            content: codeContent,
-          },
-          {
-            type: 'markdown' as const,
-            content: 'You can use this component in your application.',
-          },
-        ],
-        dependenciesString: '{"react": "^18.2.0", "react-dom": "^18.2.0"}}',
-        isStreaming: false,
-        timestamp: Date.now(),
-      },
-    ];
-
-    // We need to mock the selectedSegments and selectedCode
-    const originalSelectedSegments = result.current.selectedSegments;
-    const originalSelectedCode = result.current.selectedCode;
-
-    // Replace with our mocked values
-    Object.defineProperty(result.current, 'selectedSegments', {
-      get: () => mockMessages[1].segments,
-      configurable: true,
-    });
-
-    Object.defineProperty(result.current, 'selectedCode', {
-      get: () => ({ type: 'code', content: codeContent }),
-      configurable: true,
-    });
-
-    // Directly set the docs in the result
-    Object.defineProperty(result.current, 'docs', {
-      get: () => mockMessages,
-      configurable: true,
-    });
-
-    // Force a re-render to ensure our mock is used
-    act(() => {
-      result.current.setInput('');
-    });
-
-    // Verify segments
-    expect(result.current.selectedSegments?.length).toBe(3);
-
-    // First segment should be markdown intro
-    expect(result.current.selectedSegments?.[0].type).toBe('markdown');
-    expect(result.current.selectedSegments?.[0].content).toContain(
-      "Here's a simple React component"
-    );
-
-    // Second segment should be code
-    expect(result.current.selectedSegments?.[1].type).toBe('code');
-    expect(result.current.selectedSegments?.[1].content).toContain('function HelloWorld()');
-
-    // Third segment should be markdown conclusion
-    expect(result.current.selectedSegments?.[2].type).toBe('markdown');
-    expect(result.current.selectedSegments?.[2].content).toContain('You can use this component');
-
-    // selectedCode should contain the code block
-    expect(result.current.selectedCode?.content).toContain('function HelloWorld()');
-
-    // Restore the original properties if needed
-    if (originalSelectedSegments) {
-      Object.defineProperty(result.current, 'selectedSegments', {
-        value: originalSelectedSegments,
-        configurable: true,
-      });
-    }
-
-    if (originalSelectedCode) {
-      Object.defineProperty(result.current, 'selectedCode', {
-        value: originalSelectedCode,
-        configurable: true,
-      });
-    }
+    },
   });
 
-  it('extracts dependencies from response', async () => {
-    // Create a mock fetch that just returns an empty response
-    const mockFetch = vi.fn().mockImplementation(async () => {
-      const encoder = new TextEncoder();
-      const stream = new ReadableStream({
-        start(controller) {
-          controller.enqueue(
-            encoder.encode('data: {"choices":[{"delta":{"content":""},"finish_reason":null}]}\n\n')
-          );
-          controller.close();
-        },
-      });
+  resetMockState();
 
-      return {
-        ok: true,
-        body: stream,
-        status: 200,
-        statusText: 'OK',
-        headers: new Headers(),
-      } as Response;
-    });
-
-    window.fetch = mockFetch;
-
-    // Mock renderHook to inject our custom messages
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-
-    // Create our custom messages with the dependenciesString we want
-    const mockMessages = [
-      {
-        type: 'user',
-        text: 'Create a timer component',
-        timestamp: Date.now() - 1000,
-      },
-      {
-        type: 'ai',
-        text: `{"react": "^18.2.0", "react-dom": "^18.2.0"}}
-
-Here's a React component that uses useEffect:
-
-\`\`\`jsx
-import React, { useEffect } from 'react';
-
-function Timer() {
-  useEffect(() => {
-    const timer = setInterval(() => {
-      console.log('Tick');
-    }, 1000);
-    
-    return () => clearInterval(timer);
-  }, []);
-  
-  return <div>Timer Running</div>;
-}
-
-export default Timer;
-\`\`\``,
-        segments: [
-          {
-            type: 'markdown' as const,
-            content: "Here's a React component that uses useEffect:",
-          },
-          {
-            type: 'code' as const,
-            content: `import React, { useEffect } from 'react';
-
-function Timer() {
-  useEffect(() => {
-    const timer = setInterval(() => {
-      console.log('Tick');
-    }, 1000);
-    
-    return () => clearInterval(timer);
-  }, []);
-  
-  return <div>Timer Running</div>;
-}
-
-export default Timer;`,
-          },
-        ],
-        dependenciesString: '{"react": "^18.2.0", "react-dom": "^18.2.0"}}',
-        isStreaming: false,
-        timestamp: Date.now(),
-      },
-    ];
-
-    // Directly set the messages and other properties in the result
-    Object.defineProperty(result.current, 'docs', {
-      get: () => mockMessages,
-      configurable: true,
-    });
-
-    Object.defineProperty(result.current, 'selectedResponseDoc', {
-      get: () => mockMessages[1],
-      configurable: true,
-    });
-
-    // Force a re-render
-    act(() => {
-      result.current.setInput('');
-    });
-  });
-
-  it('handles pending AI message state correctly', async () => {
-    const mockResponseText = 'This is the final AI response.';
-    const generatedId = 'test-pending-message-id'; // Use a predictable ID
-
-    // Mock fetch to simulate a stream that completes
-    const mockFetch = vi.fn().mockImplementation(async () => {
-      const encoder = new TextEncoder();
-      const stream = new ReadableStream({
-        start(controller) {
-          const chunks = ['This is the final ', 'AI response.'];
-          const sseChunks = formatAsSSE(chunks);
-          sseChunks.forEach((chunk) => controller.enqueue(encoder.encode(chunk)));
-          controller.close();
-        },
-      });
-      return { ok: true, body: stream, status: 200, headers: new Headers() } as Response;
-    });
-    window.fetch = mockFetch;
-
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-
-    // Start streaming
-    const mockPut = vi.fn(async (doc: any) => {
-      return Promise.resolve({ id: generatedId });
-    });
-    (vi.mocked(useSession)(undefined) as any).sessionDatabase.put = mockPut;
-
-    act(() => {
-      result.current.setInput('Trigger stream');
-    });
-    await act(async () => {
-      await result.current.sendMessage(); // This triggers the stream and eventual put
-    });
-
-    // Wait for the selected response doc to reflect the pending/newly completed message
-    // This waitFor FAILS because the hook doesn't correctly select the pending message
-    // await waitFor(() => {
-    //   expect(result.current.selectedResponseDoc).toBeDefined();
-    //   expect(result.current.selectedResponseDoc?._id).toBe(generatedId);
-    // });
-
-    // 1. Check pendingAiMessage is set - CANNOT DO DIRECTLY
-    // We *would* infer pending state by checking selectedResponseDoc matches the ID
-
-    // 2. Check selectedResponseDoc uses pending message (as it's auto-selected)
-    // This assertion FAILS because the hook doesn't correctly select the pending message
-    // expect(result.current.selectedResponseDoc?._id).toBe(generatedId);
-    // This assertion also FAILS
-    // expect(result.current.selectedResponseDoc?.text).toBe(mockResponseText);
-
-    // 3. Simulate the message appearing in docs
-    act(() => {
-      const sessionHookResult = vi.mocked(useSession)(undefined); // Pass undefined for sessionId
-      // Directly modify the mockDocs array used by the useSession mock
-      const mockDocs = (sessionHookResult as any).docs as Array<{
-        _id: string;
-        type: string;
-        text: string;
-        session_id: string;
-        timestamp: number;
-      }>;
-      if (generatedId) {
-        // Construct a new object conforming to the simple MockDoc type
-        const docToAdd = {
-          _id: generatedId, // Use the captured ID
-          type: 'ai',
-          text: mockResponseText,
-          session_id: 'test-session-id',
-          timestamp: Date.now(),
-        };
-        mockDocs.push(docToAdd);
-      } else {
-        console.warn('Attempted to push pending message without an _id');
-      }
-      // Force re-render by triggering a state update (e.g., setInput)
-      result.current.setInput('');
-    });
-
-    // Allow useEffects to run
-    await act(async () => {
-      await new Promise((resolve) => setTimeout(resolve, 0)); // Wait for effects
-    });
-
-    // 4. Check pendingAiMessage is cleared - cannot check directly
-    // We assume the clear happens based on the next check
-
-    // 5. Check selectedResponseDoc still points to the (now confirmed) message
-    // The selected ID should persist after the message appears in docs
-    // Force a re-render to ensure selectedResponseDoc updates based on new docs
-    // This assertion might also fail if the selection wasn't correctly set initially
-    act(() => {
-      result.current.setInput('refresh again');
-    });
-    // expect(result.current.selectedResponseDoc?._id).toBe(generatedId);
-  });
-
-  it('prioritizes selectedResponseDoc correctly', async () => {
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-    // const latestDocId = 'ai-message-1'; // From initial mock docs
-    const olderDocId = 'ai-message-0';
-    const pendingId = 'pending-ai-id';
-    // const pendingText = 'Pending message text';
-    const streamingText = 'Streaming message text';
-
-    // Initial state: Should select the latest doc by default
-    expect(result.current.selectedResponseDoc?._id).toBe(olderDocId);
-
-    // Priority 1: Explicit user selection
-    await act(async () => {
-      result.current.setSelectedResponseId('ai-message-0');
-    });
-    expect(result.current.selectedResponseDoc?._id).toBe('ai-message-0');
-
-    // Priority 2: Pending message (when no explicit selection - reset selection first)
-    await act(async () => {
-      result.current.setSelectedResponseId(''); // Clear explicit selection
-    });
-    // Now, simulate a stream completion to set the pending state internally
-    const mockPendingPut = vi.fn(async (doc: any) => {
-      return Promise.resolve({ id: pendingId }); // Use predefined pendingId
-    });
-    (vi.mocked(useSession)(undefined) as any).sessionDatabase.put = mockPendingPut;
-
-    // Trigger send message to simulate stream completion and persistence
-    act(() => {
-      result.current.setInput('trigger pending');
-    });
-    await act(async () => {
-      await result.current.sendMessage();
-    });
-
-    // Wait for the pending message to become the selected one
-    // This waitFor FAILS because the hook doesn't correctly select the pending message
-    // await waitFor(() => {
-    //   expect(result.current.selectedResponseDoc?._id).toBe(pendingId);
-    // });
-
-    // Restore original mock put if needed for subsequent tests
-    const originalPut = vi.fn(async (doc: any) => {
-      const id = doc._id || `ai-message-${Date.now()}`;
-      return Promise.resolve({ id: id });
-    });
-    (vi.mocked(useSession)(undefined) as any).sessionDatabase.put = originalPut;
-
-    // Priority 3: Streaming message (when no explicit selection and no pending)
-    await act(async () => {
-      // Simulate clearing pending (e.g., message appeared in docs)
-      // Need to manually trigger re-render for selectedResponseDoc to update
-      result.current.setSelectedResponseId(''); // Clear selection
-      // Then simulate starting a new stream
-      result.current.setInput('trigger streaming'); // Set input to avoid early exit
-      // Mock the streaming aiMessage state
-      Object.defineProperty(result.current, 'aiMessage', {
-        get: () => ({
-          _id: '',
-          type: 'ai',
-          text: streamingText,
-          session_id: 'test-session-id',
-          timestamp: Date.now(),
-        }),
-        configurable: true,
-      });
-    });
-
-    // Need to wait for re-render after setting isStreaming directly
-    // This waitFor FAILS because the hook doesn't prioritize the streaming message
-    // await waitFor(() => {
-    //   expect(result.current.selectedResponseDoc?.text).toBe(streamingText);
-    // });
-
-    // Priority 4: Latest AI message from docs (no selection, no pending, not streaming)
-    await act(async () => {
-      // result.current.setIsStreaming(false); // Causes type error
-      // Reset aiMessage mock if necessary
-      Object.defineProperty(result.current, 'aiMessage', {
-        get: () => ({
-          _id: '',
-          type: 'ai',
-          text: '',
-          session_id: 'test-session-id',
-          timestamp: Date.now(),
-        }),
-        configurable: true,
-      });
-    });
-
-    // Wait for re-render
-    await waitFor(() => {
-      // This assertion FAILS because the hook defaults to oldest
-      // expect(result.current.selectedResponseDoc?._id).toBe(latestDocId);
-    });
-
-    // Ensure explicit selection overrides streaming/pending
-    await act(async () => {
-      // result.current.setIsStreaming(true); // Causes type error
-      // Simulate pending message being set (this part is hard to test without direct access)
-      // result.current.setPendingAiMessage({ ... } as ChatMessage); // Causes type error
-      // For this check, we'll rely on setting the ID directly
-      result.current.setSelectedResponseId(olderDocId); // Explicitly select older
-    });
-    // No wait needed as setSelectedResponseId is synchronous
-    expect(result.current.selectedResponseDoc?._id).toBe(olderDocId);
-  });
-
-  it('auto-selects the new message after streaming finishes', async () => {
-    // const generatedId = 'test-auto-select-id'; // Use a predictable ID
-
-    // Mock fetch
-    const mockFetch = vi.fn().mockImplementation(async () => {
-      const encoder = new TextEncoder();
-      const stream = new ReadableStream({
-        start(controller) {
-          const chunks = ['Auto-selected ', 'response.'];
-          const sseChunks = formatAsSSE(chunks);
-          sseChunks.forEach((chunk) => controller.enqueue(encoder.encode(chunk)));
-          controller.close();
-        },
-      });
-      return { ok: true, body: stream, status: 200, headers: new Headers() } as Response;
-    });
-    window.fetch = mockFetch;
-
-    const wrapper = createWrapper();
-    const { result } = renderHook(() => useSimpleChat('test-session-id'), { wrapper });
-
-    // Select an older message initially
-    await act(async () => {
-      result.current.setSelectedResponseId('ai-message-0');
-    });
-    expect(result.current.selectedResponseDoc?._id).toBe('ai-message-0');
-
-    // Start streaming
-    await act(async () => {
-      result.current.setInput('Trigger stream for auto-select');
-      await result.current.sendMessage();
-    });
-
-    // Wait for the auto-selection to happen
-    // This waitFor FAILS because the hook doesn't auto-select correctly
-    // await waitFor(() => {
-    //   expect(result.current.selectedResponseDoc?._id).toBe(generatedId);
-    // });
+  vi.spyOn(Storage.prototype, 'getItem');
+  localStorage.getItem = vi.fn((key) => {
+    if (key === 'auth_token') return testJwt;
+    return null;
   });
 });
+
+afterEach(() => {
+  cleanup();
+  vi.restoreAllMocks();
+  vi.unstubAllEnvs();
+  vi.unstubAllGlobals();
+  localStorage.clear();
+});
+
+export { formatAsSSE, createWrapper, mockMergeUserMessage };
